Resolves conflicts with upstream #20513 (inverted index bundle format),
#20482 (recluster sort/merge task kinds), #20498 (vacuum gc-root hashes),
#20487 (typed virtual column read slots), #20400 (streaming part reads),
#20479 (execute2 BoxFuture) and the nightly-2026-08-28 toolchain bump.

- Inverted index: carry index_version through the block index specs,
  emit BlockIndexMeta into BlockMeta.inverted_index_metas, and stream
  the new bundle format through the low-level blocking writer.
- Recluster: keep ReclusterCandidateKind (selector class) alongside the
  new ReclusterTaskKind (sort-stage work); all_ordered is replaced by
  kind == MergeBlocks.
- Vacuum: collect protected block UUIDs during the segment scan so the
  granule payload vacuum no longer needs full protected block paths.
- Virtual columns: rebuild the granule/array-cache read path on top of
  read_slots; VirtualColumnReader::read_schema is slot-keyed.
- ReadDataTransform: keep granule classification but stream one part
  per output block with runtime-filter rechecks, as upstream does.

+++ src/query/storages/common/index/src/hnsw_index/quantization/encoded_vectors_u8.rs
@@ -176,7 +176,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     }
                 };
 
-                return self.metadata.multiplier * score as f32 + query.offset + vector_offset;
+                return self.metadata.multiplier * score + query.offset + vector_offset;
             }
         }
 
@@ -192,7 +192,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     }
                 };
 
-                return self.metadata.multiplier * score as f32 + query.offset + vector_offset;
+                return self.metadata.multiplier * score + query.offset + vector_offset;
             }
         }
 
@@ -208,7 +208,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     }
                 };
 
-                return self.metadata.multiplier * score as f32 + query.offset + vector_offset;
+                return self.metadata.multiplier * score + query.offset + vector_offset;
             }
         }
